requirements.txt
astroid==1.5.3
attrs==17.4.0
autopep8==1.3.3
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorama==0.3.9
cx-Oracle==6.0rc2
defusedxml==0.5.0
Flask==0.12.2
future==0.16.0
grab==0.6.38
idna==2.5
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==3.8.0
MarkupSafe==1.0
mccabe==0.6.1
peewee==2.10.1
pep8==1.7.1
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pylint==1.7.2
PySocks==1.6.8
pytest==3.4.2
python-telegram-bot==9.0.0
pytils==0.3
requests==2.18.2
selection==0.0.13
six==1.10.0
urllib3==1.22
user-agent==0.1.9
weblib==0.1.25
Werkzeug==0.12.2
wrapt==1.10.11
